test: cover CSV whitespace and blank lines
