+++ src/components/EmployeeCombo.tsx
@@ -1,0 +1,24 @@
+import { useState, useEffect, useRef, useMemo } from "react";
+import type { Employee } from "../App";
+import { matchText } from "../utils/text";
+
+export default function EmployeeCombo({ employees, onSelect }:{ employees:Employee[]; onSelect:(id:string)=>void }){
+  const [open,setOpen]=useState(false); const [q,setQ]=useState(""); const ref=useRef<HTMLDivElement>(null);
+  const list = useMemo(()=> employees.filter(e=> matchText(q, `${e.firstName} ${e.lastName} ${e.id}`)).slice(0,50), [q,employees]);
+  useEffect(()=>{ const onDoc=(e:MouseEvent)=>{ if(!ref.current) return; if(!ref.current.contains(e.target as Node)) setOpen(false); }; document.addEventListener("mousedown", onDoc); return ()=> document.removeEventListener("mousedown", onDoc); },[]);
+  return (
+    <div className="dropdown" ref={ref}>
+      <input placeholder="Type name or ID…" value={q} onChange={e=>{ setQ(e.target.value); setOpen(true); }} onFocus={()=> setOpen(true)} />
+      {open && (
+        <div className="menu">
+          {list.map(e=> (
+            <div key={e.id} className="item" onClick={()=>{ onSelect(e.id); setQ(`${e.firstName} ${e.lastName} (${e.id})`); setOpen(false); }}>
+              {e.firstName} {e.lastName} <span className="pill" style={{marginLeft:6}}>{e.classification} {e.status}</span>
+            </div>
+          ))}
+          {!list.length && <div className="item" style={{opacity:.7}}>No matches</div>}
+        </div>
+      )}
+    </div>
+  );
+}
