+++ src/utils/csv.ts
@@ -1,0 +1,40 @@
+export function parseCSV(input: string): Record<string, string>[] {
+  const lines = input.trim().split(/\r?\n/);
+  if (lines.length === 0) return [];
+  const headers = splitLine(lines[0]);
+  const rows: Record<string, string>[] = [];
+  for (const line of lines.slice(1)) {
+    if (!line.trim()) continue;
+    const values = splitLine(line);
+    const row: Record<string, string> = {};
+    headers.forEach((h, i) => {
+      row[h] = values[i] ?? "";
+    });
+    rows.push(row);
+  }
+  return rows;
+}
+
+function splitLine(line: string): string[] {
+  const result: string[] = [];
+  let current = "";
+  let inQuotes = false;
+  for (let i = 0; i < line.length; i++) {
+    const char = line[i];
+    if (char === '"') {
+      if (inQuotes && line[i + 1] === '"') {
+        current += '"';
+        i++; // skip escaped quote
+      } else {
+        inQuotes = !inQuotes;
+      }
+    } else if (char === ',' && !inQuotes) {
+      result.push(current.trim());
+      current = "";
+    } else {
+      current += char;
+    }
+  }
+  result.push(current.trim());
+  return result;
+}
