+++ src/utils/vacancy.ts
@@ -1,0 +1,34 @@
+import type { Vacancy, Settings } from "../App";
+import { formatDateLong, combineDateTime } from "./date";
+
+export const displayVacancyLabel = (v: Vacancy) => {
+  const d = formatDateLong(v.shiftDate);
+  return `${d} • ${v.shiftStart}–${v.shiftEnd} • ${v.wing ?? ''} • ${v.classification}`.replace(/\s+•\s+$/, "");
+};
+
+export function pickWindowMinutes(v: Vacancy, settings: Settings){
+  const known = new Date(v.knownAt);
+  const shiftStart = combineDateTime(v.shiftDate, v.shiftStart);
+  const hrsUntilShift = (shiftStart.getTime() - known.getTime()) / 3_600_000;
+  if (hrsUntilShift < 2) return settings.responseWindows.lt2h;
+  if (hrsUntilShift < 4) return settings.responseWindows.h2to4;
+  if (hrsUntilShift < 24) return settings.responseWindows.h4to24;
+  if (hrsUntilShift < 72) return settings.responseWindows.h24to72;
+  return settings.responseWindows.gt72;
+}
+
+export function deadlineFor(v: Vacancy, settings: Settings){
+  const winMin = pickWindowMinutes(v, settings);
+  return new Date(new Date(v.knownAt).getTime() + winMin*60000);
+}
+
+export function fmtCountdown(msLeft: number){
+  const neg = msLeft < 0; const abs = Math.abs(msLeft);
+  const totalSec = Math.floor(abs/1000);
+  const d = Math.floor(totalSec / 86400);
+  const h = Math.floor((totalSec % 86400) / 3600);
+  const m = Math.floor((totalSec % 3600) / 60);
+  const s = totalSec % 60;
+  const core = d>0 ? `${d}d ${h}h` : h>0 ? `${h}h ${m}m` : `${m}m ${s}s`;
+  return neg ? `Due ${core} ago` : core;
+}
